requirements2.txt
joblib==1.3.2
matplotlib==3.7.4
mediapipe==0.10.9
numpy==1.24.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
pillow==10.2.0
plyfile==1.0.3
requests==2.31.0
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.1
trimesh==4.1.2
pyrender==0.1.45
smplx==0.1.14